requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
websockets==13.1
PyJWT==2.10.1
pydantic==2.10.6
requests==2.32.3
